Add files via upload
Project Objective:

The primary objective of this project is to demonstrate the potential of web scraping for collecting valuable information from online sources, especially large datasets such as the list of largest sports contracts.

Project Highlights:

Data Retrieval with Pandas: The project commences by importing the Pandas library, a versatile tool for data manipulation and analysis in Python. Pandas offers built-in functions that make web scraping and data handling seamless.

Web Scraping Wikipedia Tables: The project's focal point is the extraction of tabular data from the specified Wikipedia URL. By employing Pandas' read_html() function, the project efficiently extracts all tables present on the page.

Quantifying Data Tables: The project employs the len(tables) function to determine the number of tables extracted from the Wikipedia page. This step enables an understanding of the page's structure and how much data has been successfully retrieved.

Dataframe Creation and Exploration: The core of the project involves creating a Pandas DataFrame, termed df, by assigning the first extracted table to it. This DataFrame structures the data in a tabular format that is easily accessible and manipulatable.